[FEAT] : 아이디 검색 서버 붙이기
