+++ app/src/main/AndroidManifest.xml
@@ -19,6 +19,7 @@
             android:exported="true"
             android:windowSoftInputMode="adjustPan" />
         <activity android:name=".otherpage.view.OtherPageActivity" />
+
         <service
             android:name=".MyFirebaseMessagingService"
             android:enabled="true"
@@ -31,9 +32,13 @@
         <activity android:name=".following.view.FollowingShowAllActivity" />
         <activity android:name=".explore.view.ExploreDetailActivity" />
         <activity android:name=".answer.view.AnswerActivity" />
-        <activity android:name=".notification.view.NotificationActivity"   />
+        <activity
+            android:name=".notification.view.NotificationActivity"
+            android:exported="true" />
         <activity android:name=".detail.view.DetailActivity" />
-        <activity android:name=".idsearchfollowing.view.FollowingAfterIdSearchActivity" android:exported="true"/>
+        <activity
+            android:name=".idsearchfollowing.view.FollowingAfterIdSearchActivity"
+            android:exported="true" />
         <activity android:name=".main.view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
