+++ app/src/main/java/com/teambeme/beme/notification/model/ResponseNoticeData.kt
@@ -1,10 +1,7 @@
 package com.teambeme.beme.notification.model
 
-import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
-import kotlinx.android.parcel.Parcelize
 
-@Parcelize
 data class ResponseNoticeData(
     @SerializedName("data")
     val `data`: Data,
@@ -14,28 +11,28 @@ data class ResponseNoticeData(
     val status: Int,
     @SerializedName("success")
     val success: Boolean
-) : Parcelable {
-    @Parcelize
+) {
     data class Data(
         @SerializedName("activities")
         val activities: List<Activity>,
         @SerializedName("page_len")
         val pageLen: Int
-    ) : Parcelable {
-        @Parcelize
+    ) {
         data class Activity(
+            @SerializedName("answer_id")
+            val answerId: Int,
             @SerializedName("createdAt")
             val createdAt: String,
             @SerializedName("profile_img")
-            val profileImg: String?,
+            val profileImg: String,
             @SerializedName("question_title")
-            val questionTitle: String?,
+            val questionTitle: String,
             @SerializedName("type")
             val type: String,
             @SerializedName("user_id")
             val userId: Int,
             @SerializedName("user_nickname")
             val userNickname: String
-        ) : Parcelable
+        )
     }
 }
