+++ app/src/main/java/com/teambeme/beme/notification/view/NotificationActivity.kt
@@ -1,11 +1,8 @@
 package com.teambeme.beme.notification.view
 
 import android.os.Bundle
-import android.util.Log
 import android.view.View
 import androidx.activity.viewModels
-import com.google.android.gms.tasks.OnCompleteListener
-import com.google.firebase.messaging.FirebaseMessaging
 import com.teambeme.beme.R
 import com.teambeme.beme.base.BindingActivity
 import com.teambeme.beme.data.remote.datasource.NoticeDataSourceImpl
@@ -43,22 +40,7 @@ class NotificationActivity : BindingActivity<ActivityNotificationBinding>(R.layo
         setClickListenerForPlusData(binding)
         binding.btnBackNotice.setOnClickListener { finish() }
 
-        FirebaseMessaging.getInstance().token.addOnCompleteListener(OnCompleteListener { task ->
-            if (!task.isSuccessful) {
-                Log.w(TAG, "Fetching FCM registration token failed", task.exception)
-                return@OnCompleteListener
-            }
-
-            // Get new FCM registration token
-            val token = task.result
-
-            // Log and toast
-            val msg = getString(R.string.msg_token_fmt, token)
-            Log.d(TAG, msg)
-        })
-
-        binding.btnBackNotice.setOnClickListener {
-            finish()
+        binding.btnBackNotice.setOnClickListener { finish()
         }
     }
