+++ app/src/main/res/layout/activity_following_after_id_search.xml
@@ -115,8 +115,9 @@
                 android:id="@+id/rcv_following_after_idsearch"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
+                android:orientation="vertical"
                 android:overScrollMode="never"
-                android:visibility="@{idSearchViewModel.searchedData == null ? View.VISIBLE : View.INVISIBLE}"
+                android:visibility="visible"
                 app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
@@ -127,7 +128,7 @@
                 android:id="@+id/notice_when_no_search_data"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:visibility="@{idSearchViewModel.searchedData == null ? View.INVISIBLE : View.VISIBLE}"
+                android:visibility="invisible"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
